Re #162 Changes necessary for IO to work on Matlab 2012b
git-svn-id: https://svn.isis.rl.ac.uk/Horace/branches/OO_IO_operations@1330 1cb955da-4463-431d-88dd-2af2f5550de3

Former-commit-id: 097fe49fd4ab5155b85483456aefc96ed859402a

+++ sqw/file_io/@sqw_serializer/private/serialize_.m
@@ -44,7 +44,7 @@ for i=1:numel(fn)
             is = obj.class_map_.isKey(type);
             if is
                 nel = numel(val);
-                fnel = prod(fmt);
+                fnel = prod(double(fmt));
                 if nel ~= fnel
                     error('STRUCT_SERIALIZER:invalid_argument',...
                         'format string for constant field %s contains %d elements but field itself has %d elements',...
